backend/requirements.txt
# dependencias del backend FamilyMed

# framework
Django>=5.0
djangorestframework>=3.15

# conexion a MongoDB Atlas
pymongo>=4.6
certifi>=2024.0

# variables de entorno desde .env
python-dotenv>=1.0
